# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "法学语义学习工作台",
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/study-events/import": {
      "post": {
        "summary": "Import Study Events",
        "operationId": "import_study_events_api_study_events_import_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortableStudyEventsImport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Import Study Events Api Study Events Import Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortableStudyAttemptEvent": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "const": "retrieval_attempt",
            "title": "Event Type",
            "default": "retrieval_attempt"
          },
          "item_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Item Id"
          },
          "item_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Item Version"
          },
          "content_hash": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Content Hash"
          },
          "base_last_attempt_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Last Attempt Id"
          },
          "occurred_at": {
            "type": "string",
            "format": "date-time",
            "title": "Occurred At"
          },
          "response_text": {
            "type": "string",
            "maxLength": 10000,
            "title": "Response Text",
            "default": ""
          },
          "rating": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "again",
                  "hard",
                  "good",
                  "easy"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "elapsed_ms": {
            "type": "integer",
            "maximum": 86400000.0,
            "minimum": 0.0,
            "title": "Elapsed Ms",
            "default": 0
          },
          "revealed_answer": {
            "type": "boolean",
            "title": "Revealed Answer",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "item_id",
          "item_version",
          "content_hash",
          "occurred_at"
        ],
        "title": "PortableStudyAttemptEvent"
      },
      "PortableStudyDevice": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "label": {
            "type": "string",
            "maxLength": 120,
            "title": "Label",
            "default": "portable"
          },
          "client": {
            "type": "string",
            "maxLength": 120,
            "title": "Client",
            "default": "portable-reviewer/0.1"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "PortableStudyDevice"
      },
      "PortableStudyEventsImport": {
        "properties": {
          "protocol": {
            "type": "string",
            "const": "study-events/0.1",
            "title": "Protocol"
          },
          "bundle_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Bundle Id"
          },
          "pack_id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Pack Id"
          },
          "pack_hash": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Pack Hash"
          },
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          },
          "device": {
            "$ref": "#/components/schemas/PortableStudyDevice"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/PortableStudyAttemptEvent"
            },
            "type": "array",
            "maxItems": 1000,
            "minItems": 1,
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "protocol",
          "bundle_id",
          "pack_id",
          "pack_hash",
          "exported_at",
          "device",
          "events"
        ],
        "title": "PortableStudyEventsImport"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}